==> 1vsAll/03.GBTrainingNN.ipynb <==
import numpy as np
import pandas as pd
from pandas import DataFrame as df
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import LabelEncoder
from keras.utils import np_utils
from keras.models import Sequential
from keras.layers import Dense
from keras.layers import Dropout
import keras.regularizers as kr
from keras.wrappers.scikit_learn import KerasClassifier
from sklearn.model_selection import KFold
from sklearn.model_selection import cross_val_score
import keras.losses as ls
import matplotlib.pyplot as plt
import matplotlib
import sys
np.set_printoptions(threshold=sys.maxsize)
%matplotlib qt
font = {'family' : 'normal',
        'weight' : 'normal',
        'size'   : 14}

matplotlib.rc('font', **font)

###Target disponibili: 'NGC2243', 'NGC2420', 'Rup134', 'Trumpler20','Trumpler23'

data=pd.read_csv('./Settings/Data/Default/DEFGBData.csv')
err=pd.read_csv('./Settings/Data/DEFGBErr.csv')
nancount=[]
ind=[]
for i in range(0,data.shape[0]):
    nancount.append(data.iloc[i,12:].isnull().sum())
    if nancount[i]==0:
        ind.append(i)
    else:
        pass
data=data.iloc[ind,:]
target='Trumpler23'
targetind=[]
notargetind=[]
for i in range(0,data.shape[0]):
    if data.iloc[i,1]==target:
        targetind.append(i)
    else:
        notargetind.append(i)
tind=np.array(targetind)
notind=np.array(notargetind)
np.random.shuffle(tind)
np.random.shuffle(notind)

notvartrainind=notind[0:int(round(0.8*notind.shape[0]))]
notvarvalind=notind[int(round(0.8*notind.shape[0])):notind.shape[0]]

tvartrainind=tind[0:int(round(0.3*tind.shape[0]))]
tvarvalind=tind[int(round(0.3*tind.shape[0])):tind.shape[0]]

factor=int(notvartrainind.shape[0]/tvartrainind.shape[0])

dataset=data.values[:,12:].astype(float)
datafei=data.values[:,9].astype(float)
dataset=np.nan_to_num(dataset)
datafei=np.nan_to_num(datafei)
dataset=np.c_[dataset,datafei]
errset=err.values
errset=np.nan_to_num(errset)
tvartrain=np.zeros([tvartrainind.shape[0]*factor,err.shape[1]])

for i in range(0,tvartrainind.shape[0]):
    for j in range(0,factor):
        if j==0:
            tvartrain[i*factor+j,:]=dataset[tvartrainind[i],:]
        else:
            tvartrain[i*factor+j,:]=dataset[tvartrainind[i],:]+np.multiply(np.random.normal(scale=(1/3),size=errset.shape[1]),errset[tvartrainind[i],:])
    


asun=[6.24,7.60,6.47,6.34,3.15,3.15,4.95,3.93,5.64,5.64,7.50,7.50,4.99,6.22,2.21]
for i in range(0,dataset.shape[0]):
    for j in range(0,dataset.shape[1]-1):
        if j==10:
            pass
        else:
            dataset[i,j]=(dataset[i,j]-asun[j])-(dataset[i,10]-asun[10])
for i in range(0,tvartrain.shape[0]):
    for j in range(0,tvartrain.shape[1]-1):
        if j==10:
            pass
        else:
            tvartrain[i,j]=(tvartrain[i,j]-asun[j])-(tvartrain[i,10]-asun[10])

dataset=np.delete(dataset,10,1)
tvartrain=np.delete(tvartrain,10,1)

invartrain=np.concatenate((dataset[np.ndarray.tolist(notvartrainind),:],tvartrain))
outtrain=np.zeros(invartrain.shape[0])
outtrain[notvartrainind.shape[0]:]=1.

invarfit=np.concatenate((dataset[np.ndarray.tolist(notvarvalind),:],dataset[np.ndarray.tolist(tvarvalind),:]))
outfit=np.zeros(invarfit.shape[0])
outfit[(notvarvalind.shape[0]):]=1.
invarind=np.concatenate((np.ndarray.tolist(notvarvalind),np.ndarray.tolist(tvarvalind)))

###Viene generato il file con le label degli oggetti.

dummyout=np_utils.to_categorical(outtrain)

dummytest=np_utils.to_categorical(outfit)

###Modello di rete. La dimensione dei layer varia al variare della dimensione
###dell'input. 

model=Sequential()
model.add(Dense(invartrain.shape[1],input_dim=invartrain.shape[1],activation='selu',kernel_regularizer=kr.l1(0.01)))
model.add(Dense(round(invartrain.shape[1]*2/5),activation='selu',kernel_regularizer=kr.l1(0.04)))
model.add(Dropout(0.85))
model.add(Dense(2,activation='softsign'))
model.add(Dense(dummyout.shape[1],activation='softmax'))
model.compile(loss='categorical_crossentropy', optimizer='adam',metrics=['accuracy'])

###Training
ep=200
history=model.fit(invartrain,dummyout,validation_data=(invarfit,dummytest),epochs=ep)

###Valutazione del training fatto.

trainmse=model.evaluate(invartrain,dummyout,verbose=0)
testmse=model.evaluate(invarfit,dummytest,verbose=0)
trainmse,testmse

###Plot delle Loss

plt.close()
plt.title('Train vs Test Loss Functions')
plt.xlabel('Ages',fontsize='large')
plt.ylabel('Loss',fontsize='large')
plt.plot(history.history['loss'], label='Train set')
plt.plot(history.history['val_loss'], label='Test set')
plt.legend(fontsize='large')
plt.show()

###Plot della accuracy

plt.close()
plt.title('Accuracy',fontsize='large')
plt.xlabel('Ages',fontsize='large')
plt.ylabel('Accuracy',fontsize='large')
plt.scatter(list(range(1,ep+1)),history.history['val_accuracy'],marker='.',color='blue',facecolor='None',label='accuracy')

plt.legend(fontsize='large')
plt.show()

###Plot combinato di loss e accuracy

plt.close()
fig, (ax1, ax2) = plt.subplots(nrows=2, sharex=True)
plt.subplots_adjust(wspace=0, hspace=0)
ax1.plot(history.history['loss'], label='Training set')
ax1.plot(history.history['val_loss'], label='Validation set')
ax2.scatter(list(range(1,ep+1)),history.history['val_accuracy'],marker='.',color='blue',facecolor='None',label='Validation accuracy')
ax1.title.set_text('Training vs Validation Loss funtions & Accuracy, GB Stars')
plt.xlabel('Epochs',fontsize='large')
ax1.legend(loc='upper right')
ax2.legend(loc='lower right')
plt.show()

###Vengono salvati i weights e bias in 6 file distinti, che verranno caricati nel prossimo file:
###GBFinalNN. NB sono disponibili i weights trovati da me nella cartella FixedWeights.
path="./Settings/Weights/target/prova"
path=path.replace('target',target)
files=['/gbl0.npy','/gbl0bias.npy','/gbl1.npy','/gbl1bias.npy','/gbl2.npy','/gbl2bias.npy','/gbl3.npy','/gbl3bias.npy']
for i in range(len(files)):
    files[i]=path+files[i]

fixedw=model.get_weights()
fixedw=model.get_weights()
l0=model.layers[0].get_weights()
np.save(files[0],l0[0])
np.save(files[1],l0[1])
l1=model.layers[1].get_weights()
np.save(files[2],l1[0])
np.save(files[3],l1[1])
l2=model.layers[3].get_weights()
np.save(files[4],l2[0])
np.save(files[5],l2[1])
l3=model.layers[4].get_weights()
np.save(files[6],l3[0])
np.save(files[7],l3[1])

###Viene salvato il validation set e i dati dei plot
path2="./Settings/target/prova"
path2=path2.replace('target',target)
files2=['/vallossgb.npy','/trainlossgb.npy','/valaccuracygb.npy','/valsetinputgb.npy','/valsetoutputgb.npy','/valindgb.npy']
for i in range(len(files2)):
    files2[i]=path2+files2[i]
    
np.save(files2[0],history.history['val_loss'])
np.save(files2[1],history.history['loss'])
np.save(files2[2],history.history['val_accuracy'])
np.save(files2[3],invarfit)
np.save(files2[4],outfit)
np.save(files2[5],invarind)

#np.save(files2[],history.history['accuracy'])
